Setup for DatabaseModule refactoring
